Fix arg type

+++ runExperiment.py
@@ -79,7 +79,7 @@ parser.add_argument('--minibatch-discrimination', default=False, action='store_t
 parser.add_argument('--ideal-nmc', default=False, action='store_true', help='Whether to calculate ideal nmc')
 parser.add_argument('--optimize-features', default=False, action='store_true', help='Whether to minimize the distance between generated and real embeddings')
 parser.add_argument('--optimize-feat-epochs', type=int, default=20, help='How many epochs to run optimize-features for')
-parser.add_argument('--optimize-feat-lr', type=int, default=0.00001, help='lr for optimize-features')
+parser.add_argument('--optimize-feat-lr', type=float, default=0.00001, help='lr for optimize-features')
 
 
 args = parser.parse_args()
